+++ src/json_j.mli
@@ -109,7 +109,8 @@ type game_progress = Json_t.game_progress = {
   mutable frame_idx: int;
   steel_sole: steel_sole_progress;
   mutable by_room: (string * room_progress) list;
-  mutable keys_found: string list
+  mutable keys_found: string list;
+  mutable last_upgrade_claimed: int
 }
 
 type save_file = Json_t.save_file = {